refactor: remove unused csrf bypass
